Refactor GA Setup and Tracking to their own files
